refactor: api 파일 합치기
