--- src/utils/api/answer/getRetrieveAnswer.ts
@@ -1,8 +1,0 @@
-import api from '../axiosConfigs'
-
-export const getRetrieveAnswer = async (petitionId: string) => {
-  if (petitionId.length === 6 || petitionId === 'undefined')
-    return { status: 500, data: null }
-  const response = await api.get(`petitions/${petitionId}/answer`)
-  return response
-}
