--- src/utils/api/findPassword/postCreateVerificationCodeForPassword.ts
@@ -1,8 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postCreateVerificationCodeForPassword = async (payload: {
-  username: string
-}) => {
-  const response = await api.post('find-password/verifications', payload)
-  return response
-}
