--- src/utils/api/index.ts
@@ -1,33 +1,0 @@
-// answer
-export * from './answer/getRetrieveAnswer'
-
-// petitions
-export * from './petition/getPetitionById'
-export * from './petition/getPetitionCount'
-export * from './petition/postCreatePetition'
-export * from './petition/getAgreementCount'
-export * from './petition/getAgreements'
-export * from './petition/getStateOfAgreement'
-export * from './petition/postCreatePetition'
-export * from './petition/postAgreePetition'
-export * from './petition/getPetitionsByQuery'
-export * from './petition/getMyPetitionsByQuery'
-export * from './petition/getBestPetitionsByQuery'
-export * from './petition/getAnsweredPetitionsByQuery'
-export * from './petition/getExpiredPetitionsByQuery'
-export * from './petition/getHomeAnsweredPetitionsByQuery'
-
-// user
-export * from './user/postLogin'
-export * from './user/postLogout'
-export * from './user/postRegister'
-export * from './user/getUsersMe'
-
-// verfication
-export * from './verification/postConfirmVerificationCode'
-export * from './verification/postCreateVerificationCode'
-
-// findPassword
-export * from './findPassword/postConfirmVerificationCodeForPassword'
-export * from './findPassword/postCreateVerificationCodeForPassword'
-export * from './findPassword/putResetPassword'
