--- src/utils/api/petition/getPetitionById.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const getPetitionById = async (petitionId: string) => {
-  const response = await api.get(petitionId)
-  return response
-}
