--- src/utils/api/petition/getPetitionCount.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const getPetitionCount = async () => {
-  const response = await api.get('petitions/count')
-  return response
-}
