--- src/utils/api/petition/postAgreePetition.ts
@@ -1,11 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postAgreePetition = async (
-  petitionId: string,
-  payload: AgreePetition,
-) => {
-  if (petitionId.length === 6 || petitionId === 'undefined')
-    return { status: 500 }
-  const response = await api.post(`petitions/${petitionId}/agreements`, payload)
-  return response
-}
