--- src/utils/api/petition/postCreatePetition.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postCreatePetition = async (payload: PetitionsInput) => {
-  const response = await api.post('petitions', payload)
-  return response
-}
