--- src/utils/api/user/getUsersMe.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const getUsersMe = async () => {
-  const response = await api.get('users/me')
-  return response.status
-}
