--- src/utils/api/user/postLogout.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postLogout = async () => {
-  const response = await api.post('logout', null)
-  return response.status
-}
