--- src/utils/api/user/postRegister.ts
@@ -1,6 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postRegister = async (payload: Register) => {
-  const response = await api.post('users', payload)
-  return response
-}
