--- src/utils/api/verification/postConfirmVerificationCode.ts
@@ -1,9 +1,0 @@
-import api from '../axiosConfigs'
-
-export const postConfirmVerificationCode = async (payload: {
-  username: string
-  verificationCode: string
-}) => {
-  const response = await api.post('sign-up/confirm', payload)
-  return response
-}
